gen_doc: add barc commands to generated man texts
Make it easier to access the barc per-command documentation by
maintaining a copy in the repo, this also means it's published anywhere
the source code is visible.

It's important that the documentation is usable, by re-generating in
the pre-commit scripts we ensure that the output version is reviewed
alongside the code that changes it.

Test Plan:
Review the doc after running
./gen_doc.sh