Add progress indicators to test scripts
The './precommit.sh' tests are starting to take a while, make it easier
to see what's going on by displaying a sort of progress indicator.

e.g.
    gen_doc: . ....... ...OK

In the event of failure, this also makes it easier to tell which of
the steps failed - by counting the pips and comparing to the script.

Test Plan:
$ ./precommit.sh

Determined which static tester was complaining in a particular case, to
look for instructions to ignore the individual instance of the error.