abdt_arcydreporter: less log_system_error() detail
It's easily possible for 'detail' to exceed the length of command-line
parameters allowed when calling out to a registered external system
error logger.

Limit the amount of detail that will be sent to an arbitrary small
number to prevent errors when reporting errors.

Test Plan:
$ ./precommit.sh

Modify abdi_processrepo to raise a massive exception, verify it fails to
report the exception before the change and succeed after the change.
$ ./testbed/arcyd/test_shell.sh