Arcyd: Enforce config name conventions
The naming convention for config files is advertised in the help for
functions but hasn't been enforced, until now.

Especially interesting is the common '.git' in the repo url, which is
invalid under the scheme, in later work we can update the add
subcommands to remove .git from the 'guessed' config names for
convenience.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/test_shell.sh
    Try adding various invalid names
    $ cd ../arcyd
    $ arcyd add-repo localhost fs1 dev --name blah.git
    $ arcyd add-repo localhost fs1 dev --name blah-blah
    $ arcyd add-repo localhost fs1 dev --name Blah-Blah
    $ cp -R ../dev ../dev3.git
    $ arcyd add-repo localhost fs1 dev3.git