Add 'arcyon comment-inline' command
Make it possible to create an inline comment on a differential review
with a new command 'arcyon comment-inline'.

Note that a subsequent call to 'arcyon comment --attach-inlines' is
necessary to actually commit the inline comments on the review.

Test Plan:
$ ./precommit.sh

Add code to exercise inlines in ./testbed/arcyon/exercise_arcyon.sh,
verified results in browser.