Refactor, initialize ReviewStateCache with conduit
Reduce overall complexity by removing the conduit setter and getter on
phlcon_reviewstatecache.ReviewStateCache and instead requiring a conduit
when it is created.

This makes it easier to work with and reduces the testing burden.

It appears the need for updating the conduit has disappeared.

Test Plan:
Update docs, exercise changed code
$ ./precommit.sh