Fix fwd-incompatability with 'arcyon task-query'
In the current version of Phabricator, the 'status' attribute of
Maniphest tasks is now a string and is returned as such via the Conduit
API.

This breaks our conversion of int id -> statusName when trying to be
backwards compatible with older versions of Phabricator that don't fill
in 'statusName' in the Conduit response.

Work around this incompatability by changing 'statusName' to default to
'None' and only attempt to fill in the 'statusName' attribute where it
hasn't been filled in.

The change to 'status' happened after the introduction of 'statusName'
so this should always work.

Closes https://github.com/bloomberg/phabricator-tools/issues/11

Test Plan:

Exercise old workaround with old version of Phabricator
    $ arcyon task-query --uri http://127.0.0.1  --ids 6
    .. ok ..

Exercise new workaround with new version of Phabricator
    $ arcyon task-query --uri https://secure.phabricator.com  --ids 6144
    .. ok ..

Usual checks just in case
$ ./precommit.sh