Add '--external-report-command' to arcyd
Make it easier to hook up arcyd to external monitoring by providing an
optional callback to an external command, which will be provided with a
report in json.

Set this up in the test shell for easy exercising.

Test Plan:

Check that happy path logs json reports to file
$ ./testbed/arcyd/test_shell.sh
    $ tail -f ../arcyd/savereport.txt
    .. logs json reports as expected ..
    $ exit

Add 'false' line into 'savereport', to exercise unhappy case
$ ./testbed/arcyd/test_shell.sh
    .. wait ..
    $ exit
    .. error appears in logs as expected ..

Make 'savereport' non-executable, to check that we fail fast
$ chmod -x testbed/arcyd/savereport
$ ./testbed/arcyd/test_shell.sh
    .. wait ..
    $ exit
    .. fatal error appears in logs as expected ..

Check that other code paths are unaffected
$ ./precommit.sh