arcyd fsck: check specific list of repos
Make it possible to check a partial list of repositories by adding a new
parameter to support this.

This is useful if a single repo needs to be checked or to enable a
workflow where errors are listed first and fixed individually.

Test Plan:
$ ./precommit.sh

Make sure list of repos checked is as expected
$ ./testbed/arcyd/test_shell.sh
    $ cd ../arcyd2
    $ arcyd fsck -v
    .. lists repos checked ..
    $ arcyd fsck -v --repos <<valid repo>>
    .. lists single repo checked ..
    $ arcyd fsck -v --repos <<valid repo 1>> <<valid repo 2>>
    .. lists both repos checked ..
    $ arcyd fsck -v --repos
    .. error message ..
    $ arcyd fsck -v --repos <<INVALID repo>>
    .. error message ..