phlgit_fetch: add support for fetching with spec
It can be useful sometimes to customize what gets fetched from a remote,
e.g. when fetching the 'refs/notes' namespace locally.

Supplying a 'refspec' directly to 'git fetch' is one of the ways to
achieve this end.

Add tests to verify that fetching a non-existant namespace is ok.

Test Plan:
$ ./precommit.sh