Arcyd: BREAK, removed delay/retry mechanism
Pave the way towards multi-threaded arcyd by inlining all of the
'scheduled unreliable operations' into abdi_processrepolist. It turns
out that the necessary code is much smaller and simpler without all the
abstractions.

Remove the feature of delayed retying of failed repositories with the
intention of adding it back in later.

Test Plan:
Generally exercise things
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh
$ ./testbed/arcyd/unreliable_git_loop_test.sh
$ ./testbed/arcyd/test_shell.sh
    $ ./poke.sh
    $ ./accept.sh