Arcyd: save snoop cache once per cycle only
Avoid consistency concerns when saving the git snoop url cache by
hoisting it out of the multithreaded section.

Note that other usage of the cache is thread-safe as consistency only
has to be preserved per-URL. This is ensured because we'll only update a
particular repo on one thread at a time.

Remove the snoop cache sharing concern from the list of multi thread
concerns.

Test Plan:
$ ./precommit.sh