arcyd-tester: refactor, _happy_path as a generator
Convert _happy_path() to a generator so that it will be easy to write
functions that change the behaviour of Arcyd and run _happy_path() over
a number of different Arcyd setups.

This means that we will be able to separate 'user scenarios' from
'Arcyd scenarios' and write both naturally as generators.

For example, we could combine an Arcyd scenario of migrating from one
Arcyd instance to another with the user scenario of the 'happy path'.

This is more natural to achieve with generators rather than classes
because we can write each scenario as a single function rather than
perhaps making a state machine.

This will also make it easy to run a number of _happy_path() and other
user scenarios at the same time to ensure they don't interfere with
eachother.

Test Plan:
$ ./precommit.sh