abdi_processrepoarglist: retry failed repositories
Provide an allowance for temporary failures by retrying repositories
that fail in an unexpected way after a delay.

Increase the delay on each subsequent failure to limit the possible
damage from a serious failure.

Eventually stop retrying, so that any possible side-effects from the
failure have a limit.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/test_shell.sh
    .. reduce the retry delays to seconds ..
    $ tail -f ../arcyd/var/log/info &
    $ mv ../origin2 ../origin3
    .. observe retry loop failure warnings ..
    .. observe top-level exception error with delay ..
    .. arcyd retries failed repo again after a delay ..
    .. observe retry loop failure warnings ..
    .. observe top-level exception error with delay ..
    .. and so on, until the repo is no longer retried ..