Arcyd: log io read events to debug log instead
Log any reported io-read-events to the debug log instead of a special
separate file.

The 'abdt_fs.Layout.log_remote_io_reads' file quickly becomes large and
doesn't seem to have very valuable information. It's more suited to be
collected with other events of temporary interest in the debug log.

The debug log is rotated, so there's a safe upper limit on the amount of
space we'll use.

Make sure each line has 'io-read-event' in it, so it's easy to use grep
to retreive just the read events.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/test_shell.sh
    $ tail ../arcyd/var/log/debug -f
    .. observe read events being logged appropriately ..