autofix/fixlegal: consider staged files
When fixing up legal notices, esp. the dates, consider files that have
changes staged in the index too.

The most common use for this script is when running it as part of the
'./precommit.sh' script. Often there will therefore be changes staged in
the INDEX but not yet committed.

Test Plan:
.. checkout new branch, edit file not changed this year, add file ..
$ ./autofix.sh
.. copyright is updated in file with current year ..