Arcyd: debug log the duration of cycle events
Make it easier to tell what's taking time in a cycle by logging the
duration of 'process-repos', 'external-report-command' and 'sleep' to
the debug log.

Introduce abdt_logging.misc_operation_event_context() for this sort of
thing. Log these events as 'misc-event' so that they may be grepped out
of the log easily. Don't choose 'event' as that is a substring of other
kinds of events.

Test Plan:
$ ./precommit.sh
$ ./tested/arcyd/test_shell.sh
    $ tail -f ../arcyd/var/log/debug
    .. observe the new events fly past ..