phlcon_differential: handle reviewers as title
Handle the case where a commit message as the reviewers in the title,
this appears to generate a strange response for the version of
Phabricator we're building against.

Raise a specific error 'UnknownParseCommitMessageResponseError' in the
event of this.

Later versions may have a fix for this, in which case we'll want to make
sure we either raise that error or no error.

Add a test which evokes this response.

Test Plan:
$ ./precommit.sh