abdi_processrepoarglist: fix datetime comparison
We can't compare datetime objects to None, add a check for
_reactivate_time being None prior to comparing.

This fixes the case where a repository fails all retries and
subsequently throws. This was resulting in jobs to appear as
'overrunning' when they in fact only didn't report the results.

In later work we'll want to change phlmp_cyclingpool to handle raised
exceptions gracefully.

Test Plan:
$ ./precommit.sh

.. change repositories to immediately fail ..
$ ./testbed/arcyd/test_shell.sh
    $ tail -f ../arcyd/var/log/debug
    .. observe that jobs don't time out when all retries exhausted ..