Refactor, fixup new flake8 issues
A new version of the pep8 utility has been released which rightly
complains if there are module-level imports that are not at the top of a
file. This is true as of at least `pep8-1.6.1`.

Ignore this error in the strictly limited places where we do this, which
is after manipulating the system path so we can subsequently import bits
from the phabricator-tools library in launcher scripts.

Ultimately this repo should behave as a regular python package and so
would not need this special treatment.

Test Plan:
.. upgrade pep8 package ..
$ ./precommit.sh