vagrant/puppet: use newer phabricator commits
Upgrade Phabricator to a version with a fix for getting diffs from
revisions which don't have arcanist projects:
https://github.com/phacility/phabricator/commit/119c2b8

Roll Phabricator forward just a little more to also pickup a fix for
Maniphest, this was failing when trying to create a new issue:
https://github.com/phacility/phabricator/commit/960da82

Roll it on again just a bit to pick up another fix for Maniphest.
https://github.com/phacility/phabricator/commit/d13a322

libphutil didn't have a related change at the time of the new
Phabricator commits.

Arcanist was quite a long way out of date, roll it forward to roughly
the same time.

Test Plan:
.. create new ubuntu 14.04 box, provision with puppet ..
.. setup git, arcanist ..
$ ./precommit.sh
$ ./testbed/arcyd/acceptance_tests.sh