arc[yd,yon].generated.txt: update for new argparse
It appears that new versions of CPython have an implementation of
argparse which has a very slightly different output. Commit the latest
version of the output.

This generation of docs is always going to be sensitive to such changes,
if they start coming up often then we'll have to reconsider this being
part of `precommit.sh`.

Tested with `Python 2.7.6`.

Test Plan:
$ ./precommit.sh