pep8: refactor, binary operators before line break
The pep8 checker seemed to gain a new ability in 1.6.2, it now checks
for binary operators broken across lines.

For clarity the binary operator should appear before the line break
rather than after it.  Otherwise perhaps it's easier to miss that the
statement is continued on the next line.

From PEP8:

    Make sure to indent the continued line appropriately. The preferred
    place to break around a binary operator is after the operator, not
    before it.

https://www.python.org/dev/peps/pep-0008/#maximum-line-length

Fix violations detected by the pep8 tool.

Test Plan:
.. update pep8 tool ..
$ ./precommit.sh