vagrant/puppet/../default.pp: install wget
This common utility may not come pre-installed with some flavors of
ubuntu (e.g. in Docker's official image). The puppet manifest depends on
it, so explicitly install it in the puppet manifest.

Test Plan:
Create a new instance of Phabricator inside a fresh Ubuntu 14.04 Docker
container, without needing to manually install wget.