Arcyd: add file handler to compress rotated debug logs
For a large arcyd installation, the var/log/debug log file can rotate
completely multiple times per day. This means that it's not possible to
look into events from previous days. So, gzip the rotated logs so that
they take less space. In future, we should be able to increase the
amount of logs stored without increasing the size of those but that will
be done as part of a separate issue.

Testplan:
$ ./precommit.sh

Fixes #18