Meta: Fix meta/autofix/fixlegal.py to work with new files
Currently when invoked on new files, meta/autofix/fixlegal.py fails with an
exception. This happens because it tries to get last commit from commit
log which is empty if the file is new. Fix this behavior by using the
current year when git log is empty.

Test plan:
Create a new file, stage it for commit and run precommit.sh

Fixes #25