Arcyd: Increase the size of individual debug logs to 50 MB
Increase the value of maxBytes argument for the file handler to 50 MB which
is currently set to 10 MB.

The (maximum) total size of the logs would be ~100 MB (50 + (50 * 10) / 10)
assuming a gzip compression factor 10. On the other hand, current (maximum)
total size of the logs is 110 MB (10 + (10 * 10)). So we'll be able to store
5 times more logs without actually increasing the total size of the logs
stored.

Test plan:
$ ./precommit.sh

(Manual) Let Arcyd put huge amounts of stuff in debug logs and ensure that
logs are rotated when the filesize reaches 50 MB. One simple way of
doing this is to add to add following 3 lines in the start_arcyd()
function of py/abd/abdi_startstop.py just below line 83:
    while True:
        _LOGGER.debug("This is a test entry for the debug log." * 100)
        time.sleep(0.01)

With this change, the debug log should cross the 50 MB in a minute or
two. Once it does that, verify that that a debug.1.gz is created and
that the debug file is almost empty now.

Fixes #35