Arcyd-tester: Add option to disable auto debug shell
At present, if any of the tests fail, debug shell is launched
automatically. This behavior is not natural, especially for new users.
This commit adds a command line option to the script to enable the
automatic launch of debug shell but it is set to false by default. To
enable it, the script should be passed a '--enable-debug-shell' option.

Test plan:
$ ./precommit.sh

(Manual) Try to run the script with '--enable-debug-shell' and esnure
that it automatically opens the debug shell on a failure. Running it
without that option should not result in opening of a shell.