Arcyd: Fix cache race condition when overrunning
Arcyd worker processes which overran could incorrectly consume the
'newness' of a repository which had been updated while it was
overrunning. This is because it would happen unconditionally, whether or
not the repository snoop url now pointed to different content.

Fix this by using function 'watcher.merge_data_consume_only()' which
merges the data for URLs and also consumes the 'newness' of the URL only
if the content hashes match.

This commit also fixes the fail in Arcyd tester which was introduced in
a previous commit.

Test plan:
$ ./precommit.sh