Arcyd-tester: Test pushing during overrun
Add  _test_push_during_overrun() to test pushing during overrun. This
commit just adds the test for this functionality to point out that there
is some issue with existing code. So the test is expected to fail at
this stage. The fix for this issue will be added in a later commit which
will fix the issue as well as the tests.

Test plan:
$ ./precommit.sh
(./proto/arcyd-tester should fail at this point of time because just the
test is introduced in this commit and the actual fix will be added as a
separate commit.)