Meta: Add exclude option to precommit.sh
Add a '-e' option to precommit.sh to allow exclusion of certain tests
from the script. The names of the tests to be excluded are provided as
arguments to this option and valid options are 'gen_docs', 'autofix',
'static_tests', 'unit_tests' and 'smoke_tests'.

Test plan:
$ ./precommit.sh

$ ./precommit.sh -h
Ensure that correct and meaningful help text is displayed.

$ ./precommit.sh -e smoke_tests
Ensure that smoke_tests.sh is not ran when its passed as an argument to
'-e'.