phlsys_conduit: Rename make_phab_test_conduit
Rename the function make_phab_test_conduit() to
make_phab_example_conduit() because any function which has a 'test' in
its name is considered as a test by Nose. So, Nose considered this
function when it was actually just a helper function which would cause
the tests to fail in some cases. So, rename this function to avoid any
confusion.

Test plan:
$ ./precommit.sh

$ git grep make_phab_test_conduit
Ensure output of grep is empty to make sure that the old function name
is not used in any place.