Meta: Remove integration doc test
There's a doctest which in phlsys_conduit.py which is in fact an
integration test and requires a local Phabricator instance to be running
in order to pass successfully. Since we plan to keep unit tests separate
from integration tests, we do not allow doc tests to be dependent on a
Phabricator instance.

Since this is a valuable usage example, keep it as an example but do not
run it as a doc test.

Test plan:
$ ./precommit.sh

Run the following command without a Phabricator instance.
$ nosetests py/phl/phlsys_conduit.py --with-doctest
Ran 1 test in 0.002s
OK