phlurl_watcher: add ability to merge watchers
In the case where we have one process forked from another to do work,
it's useful to be able to 'consume the newness' from the forked watcher
when joining / finishing the worker process.

The main process and the worker process may have diverged in their view
of the watcher urls. In this case we will only want to 'consume the
newness' for urls where the hash of the content matches. In cases where
the hashes don't match, we can't be sure which newness is newer.

Add get_data_for_merging() and merge_data_consume_only() to facilitate
this data flow.

Add tests for this new functionality.

Test Plan:
$ ./precommit.sh