Arcyd-tester: Add test for number of fetches
Add tests to ensure that Arcyd fetches from the repository exactly as
many times as is absolutely required. We need to ensure that Arcyd does
not makes any extra fetches in order to not over-burden git server.
Counting the number of pushes is achieved by monkey patching the
base-level git clone to update the count in a file on every fetch.

Test plan:
$ ./precommit.sh