Meta: Add script to automatically generate test plan
Add a script to automatically generate test plan block at the top of
each test fixture. Since there is a lot of discrepency in the way we
describe our test plans currently, this script is not included in
precommit.sh script but it can be run for the scripts which follow all
the standards correctly.

There is also a change in the way concerns are described is the way we
prefix them. Earlier, the comments for concerns had to be prefixed like
so:
    # [ A] description...
Now, those can also be described like so:
    # CONCERN: description...

The latter version is preferred because it offers higher portability
while renaming functions and the both version are converted to the same
string in test plan description.

Test plan:
$ ./precommit.sh

Change the concern comments in py/phl/phlsys_verboseerrorfilter__t.py to
follow the new version, i.e. prefixed with "CONCERN:" and delete the
test plan from it. Run the script to update test plan like so:
    $ ./gen_test_plan.sh py/phl/phlsys_verboseerrorfilter__t.py
Manually verify that the test plan is generated correctly.