poxcmd_conduitproxy.py: disable pylint error
Newer versions of Pylint has a false positive for no-value-for-parameter
when we're passing through received parameters in _RequestHandler.

E:138, 8: No value for argument 'request' in unbound method call
(no-value-for-parameter)
E:138, 8: No value for argument 'client_address' in unbound method call
(no-value-for-parameter)
E:138, 8: No value for argument 'server' in unbound method call
(no-value-for-parameter)

Disable this error for the offending line.

Test Plan:
$ ./precommit.sh