Add docker/arcyd-cluster for experiments
When running Arcyd in production, we won't want to nervously watch a
single machine to make sure it's still running.

Provide a proof-of-concept of Arcyd running with Consul to provide
leadership election with basic status monitoring. If the Arcyd leader
stops responding then automatically elect a new leader and continue.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/docker-cluster-test.sh
.. check that only one instance becomes the leader ..