docker/arcyd/entrypoint: docker stop -> arcyd stop
Run 'arcyd stop' if the user calls 'docker stop' on the container.

'docker stop' sends SIGTERM to the pid1 of the container. Handle this
signal by calling 'arcyd stop'.

This results in cleaner shutdown behavior.

Note that 'docker stop' will send SIGKILL after 10 seconds by default if
the container has not exited, it may be wise to choose a longer timeout
for instances that take a long time to stop. This note is correct for
Docker version 1.8.2.

Test Plan:
$ ./precommit.sh
$ ./testbed/arcyd/docker-testshell.sh
    $ docker stop arcydtest-arcyd
    $ docker start arcydtest-arcyd
    $ docker exec arcydtest-arcyd cat /var/arcyd/var/log/info
    .. arcyd stops ..
    reason given: SIGTERM / probably docker stop
    .. arcyd starts ..