testbed/phlsys_fs_atomic: try kill -9 when writing
Experiment with phlsys_fs_atomic.open_write_text_file_atomic() to see if
we can successfully usually prevent a file from being partially written
if it is killed whilst attempting to write.

Test Plan:

Run tests on an ext4 filesystem in a Lubuntu 14.04.3 virtual machine.

$ cd testbed/phlsys_fs_atomic

Run the version that attempts to update the file atomically. See that we
don't catch it writing partial contents.

    $ time ./test_file_writer --atomic
    ___.__________.._.__..______....__  (.. snip ..)
        631 _
        369 .

    real    1m5.887s
    user    0m49.124s
    sys 0m7.056s

Run the version that does not attempt to update the file atomically, see
that it occasionally writes partial contents. This gives us more
confidence that the first test is valid.

    $ time ./test_file_writer
    ___.__._.____X_X._._X_________XX_X  (.. snip ..)
        637 _
        95 .
        268 X

    real    1m0.785s
    user    0m47.172s
    sys 0m6.972s