phlsys_signal: add no_exit_context, tests
In some circumstances, we will want to defer raising SystemExit due to
SIGTERM so that we can maintain consistency.

For example, when landing a review with Arcyd we want to push branches
and update Phabricator to say we've done it. If we can avoid
interruption between those steps then it's much better.

Introduce a 'no_exit_context' context manager that will defer raising
SystemExit until the context is finished. Block other exceptions from
raising through if we've already received SIGTERM.

Add tests to exercise this functionality.

Test Plan:
$ ./precommit.sh