testbed/arcyon/exercise_arcyon.sh: bake in creds
We're already assuming that we're using the database dump provided by
phabricator-tools, by relying on 'alice', 'bob', and 'phab' users being
present.

For convenience, go a step further and assume that the credentials for
the phab user are unchanged. This removes the need for folks to install
an ~/.arcrc file in order to run these tests.

Test Plan:
$ ./precommit.sh